NGOVANQUANG_20183813: modified UserModel and RoleModel changed variables primetype to objecttype

+++ Final/Backend/QuanLyKTX/src/main/java/web/nhom8/quanlyktx/dao/MysqlDao.java
@@ -53,7 +53,7 @@ public class MysqlDao {
             resultSet = statement.executeQuery();
             while (resultSet.next()) {
                 RoleModel roleModel = new RoleModel();
-                roleModel.setRoleId(resultSet.getInt("RoleId"));
+                roleModel.setRoleId(resultSet.getLong("RoleId"));
                 roleModel.setRoleName(resultSet.getString("RoleName"));
                 roleModel.setRoleCode(resultSet.getString("RoleCode"));
                 results.add(roleModel);
